+++ install.sh
@@ -5,7 +5,8 @@
 ## Please adapt this to fit your distro needs
 
 mkdir /etc/osync
-cp ./sync.conf /etc/osync
+cp ./sync.conf /etc/osync/sync.conf.example
+cp ./exclude.list.example /etc/osync
 cp ./osync.sh /usr/local/bin
 cp ./osync-srv /etc/init.d